requirements.txt
pandas>=1.3.0
requests>=2.28.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
pyyaml>=6.0
flask>=3.0.0
openpyxl>=3.1.0
matplotlib>=3.7.0
pytest>=7.0.0
